# Update user used by `Kibana` in the cluster performance tests